Removing unneeded commented-out code
