+++ src/app/types/ai.ts
@@ -37,23 +37,6 @@ export interface AIProvider {
     processingTime: number;
   }
 
-  // export interface PersonalityResponse {
-  //   personality: Personality;
-  //   alignment?: AlignmentAnalysis;
-  //   safety?: {
-  //     overallScore: number;
-  //     dimensions: Record<string, { score: number; passes: string[]; concerns: string[] }>;
-  //   } | null; // Allow null
-  //   trust?: {
-  //     score: number;
-  //     weights: { alignment: number; safety: number };
-  //     criteria?: { name: string; passed: boolean; details: string[]; explanation: string };
-  //   } | null; // Allow null
-  //   providerName: string;
-  //   processingTime: number;
-  //   response: string;
-  // }
-
   export interface PredefinedValue {
     name: string;
     principles: string[];
@@ -72,22 +55,6 @@ export interface AIProvider {
     updateHofstedeDimension: (dimension: keyof HofstedeDimensions, value: number) => void;
     removePrinciple: (index: number) => void;
   }
-  
-  // export interface MultiProviderAI {
-  //   providers: AIProvider[];
-  //   apiKeys: APIKeyConfig;
-  //   selectedModels: SelectedModel[];
-  //   error?: string;
-  //   isConfigured: () => boolean;
-  //   getConfiguredProviders: () => AIProvider[];
-  //   getSelectedModelCount: () => number;
-  //   getSelectedModels: () => SelectedModel[];
-  //   getProviderName: (providerId: string) => string;
-  //   generateSingleResponse: (prompt: string, modelId: string, providerId: string) => Promise<string>;
-  //   updateAPIKeys: (keys: APIKeyConfig) => void;
-  //   updateSelectedModels: (models: SelectedModel[]) => void;
-  //   addCustomModel: (model: AIModel) => void;
-  // }
 
   export interface MultiProviderAI {
     providers: AIProvider[];
